+++ sdk/lib/conutils/pager.c
@@ -655,7 +655,7 @@ ConPutsPaging(
     if (szStr == NULL)
         return TRUE;
 
-    len = wcslen(szStr);
+    len = (DWORD)wcslen(szStr);
     return ConWritePaging(Pager, PagePrompt, StartPaging, szStr, len);
 }
